fix(core): switch radial_normalization_factor to use u64 factorial This temporarily prevents overflows for reasonable-ish values of n and l (ex. 8, 5) but will will require a more permanent fix later in the form of expanding the two factorials and computing the ratio directly as the product of leftover terms.

+++ evanescence_core/src/orbital/wavefunction.rs
@@ -15,8 +15,8 @@ pub trait Wavefunction {
 
 #[inline]
 fn radial_normalization_factor(n: u32, l: u32) -> f64 {
-    let root_numerator = (n - l - 1).factorial();
-    let root_denominator = (n + l).factorial();
+    let root_numerator = ((n - l - 1) as u64).factorial();
+    let root_denominator = ((n + l) as u64).factorial();
     2.0 / (n * n) as f64 * (root_numerator as f64 / root_denominator as f64).sqrt()
 }
